enable temp file deletion in pack.t on Win32
The code used in sub script_command_file to slurp file contents leaves
ARGV as an open filehandle to the file last slurped. For pack.t this means
it cannot remove the temp files on windows due to the left-over filehandle.
Closing ARGV explicitly after each slurp fixes this issue.